Fix update_visibility_values

+++ Bugzilla/Field.pm
@@ -1017,11 +1017,14 @@ sub update_visibility_values
     }
     my $del = [ keys %$h ];
     return 0 if !@$add && !@$del;
-    Bugzilla->dbh->do(
-        "DELETE FROM fieldvaluecontrol WHERE field_id=? AND value_id=?".
-        " AND visibility_value_id IN (".join(", ", @$del).")",
-        undef, $self->id, $controlled_value_id
-    );
+    if (@$del)
+    {
+        Bugzilla->dbh->do(
+            "DELETE FROM fieldvaluecontrol WHERE field_id=? AND value_id=?".
+            " AND visibility_value_id IN (".join(", ", @$del).")",
+            undef, $self->id, $controlled_value_id
+        );
+    }
     $self->add_visibility_values($controlled_value_id, $add);
     return 1;
 }
@@ -1030,6 +1033,7 @@ sub add_visibility_values
 {
     my $self = shift;
     my ($controlled_value_id, $visibility_value_ids) = @_;
+    return 0 if !@$visibility_value_ids;
     my $f = $self->id;
     for ($controlled_value_id)
     {
